// src/components/today/TodayItem/type.ts
export type ItemType = {
  title: string;
  adult?: boolean;
  author: string;
  bestDuration?: string;
  bestRank: number;
  categoryId?: number;
  categoryName: string;
  cover: string;
  customerReviewRank?: number;
  description: string;
  fixedPrice?: boolean;
  isbn: string;
  isbn13?: string;
  itemId?: number;
  link?: string;
  mallType?: string;
  mileage?: number;
  priceSales?: number;
  priceStandard?: number;
  pubDate: string;
  publisher: string;
  salesPoint?: number;
  stockStatus?: string;
  subInfo?: any;
};